backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
django-cors-headers==4.3.1
drf-yasg==1.21.7
django-filter==24.2
setuptools==69.5.1
